deployment/notification_sender/requirements.txt
# requests
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
urllib3==2.5.0
# python-dateutil
python-dateutil==2.9.0.post0
six==1.17.0
